More Fixes
Fixed some more things. Still a few problems, but I’m done with it for
now. The tabbed view requires an entry point (not sure how to add one).
There is also a reference “bridge”, I’m not at all sure what it’s
supposed to be, so I left it.